requirements.txt
pydantic>=1.10.2
Flask<3.0.0
Flask-WTF
flask-sqlalchemy
Flask-Login
psycopg2-binary
python-dotenv
werkzeug<2.3
waitress
paste